+++ app/src/main/java/com/maurodev/chronometer/ui/splash/SplashScreenActivity.kt
@@ -1,0 +1,27 @@
+package com.maurodev.chronometer.ui.splash
+
+import android.annotation.SuppressLint
+import android.content.Intent
+import androidx.appcompat.app.AppCompatActivity
+import android.os.Bundle
+import android.os.Handler
+import android.os.Looper
+import com.maurodev.chronometer.MainActivity
+import com.maurodev.chronometer.databinding.ActivitySplashScreenBinding
+
+@SuppressLint("CustomSplashScreen")
+class SplashScreenActivity : AppCompatActivity() {
+
+    private lateinit var _binding: ActivitySplashScreenBinding
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+        _binding = ActivitySplashScreenBinding.inflate(layoutInflater)
+        setContentView(_binding.root)
+
+        Handler(Looper.getMainLooper()).postDelayed({
+            val intent = Intent(this, MainActivity::class.java)
+            startActivity(intent)
+            finish()
+        }, 3000)
+    }
+}