+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Fecha XX/XX/XX
 
 ### Nuevas Características
+- Implementacion del splash screen.
 
 ## Fecha 13/08/23